refactor(update-repo): modularize utility functions
Extract helper functions for section parsing and chezmoi templating into dedicated modules. This improves code organization and separation of concerns for the update_repo script.
